Extracted out handles into their own component and improved hitbox
