+++ src/renderer/components/inputs/InputContainer.tsx
@@ -1,64 +1,16 @@
 import { Type } from '@chainner/navi';
-import { Box, Center, HStack, Text, Tooltip, chakra } from '@chakra-ui/react';
+import { Box, Center, HStack, Text } from '@chakra-ui/react';
 import React, { memo, useCallback, useMemo } from 'react';
-import ReactMarkdown from 'react-markdown';
-import { Connection, Handle, Node, Position, useReactFlow } from 'reactflow';
+import { Connection, Node, useReactFlow } from 'reactflow';
 import { useContext } from 'use-context-selector';
 import { InputId, NodeData } from '../../../common/common-types';
 import { parseSourceHandle, parseTargetHandle, stringifyTargetHandle } from '../../../common/util';
-import { VALID, Validity, invalid } from '../../../common/Validity';
+import { VALID, invalid } from '../../../common/Validity';
 import { BackendContext } from '../../contexts/BackendContext';
 import { GlobalVolatileContext } from '../../contexts/GlobalNodeState';
 import { defaultColor, getTypeAccentColors } from '../../helpers/accentColors';
-import { noContextMenu } from '../../hooks/useContextMenu';
+import { Handle } from '../Handle';
 import { TypeTag } from '../TypeTag';
-
-interface LeftHandleProps {
-    isValidConnection: (connection: Readonly<Connection>) => boolean;
-    validity: Validity;
-}
-
-// Had to do this garbage to prevent chakra from clashing the position prop
-const LeftHandle = memo(
-    ({
-        children,
-        isValidConnection,
-        validity,
-        ...props
-    }: React.PropsWithChildren<LeftHandleProps>) => (
-        <Tooltip
-            hasArrow
-            borderRadius={8}
-            display={validity.isValid ? 'none' : 'block'}
-            label={
-                validity.isValid ? undefined : (
-                    <ReactMarkdown>{`Unable to connect: ${validity.reason}`}</ReactMarkdown>
-                )
-            }
-            mt={1}
-            opacity={validity.isValid ? 0 : 1}
-            openDelay={500}
-            px={2}
-            py={1}
-        >
-            <Handle
-                isConnectable
-                className="input-handle"
-                isValidConnection={isValidConnection}
-                position={Position.Left}
-                type="target"
-                // eslint-disable-next-line react/jsx-props-no-spreading
-                {...props}
-            >
-                {children}
-            </Handle>
-        </Tooltip>
-    )
-);
-
-const Div = chakra('div', {
-    baseStyle: {},
-});
 
 export interface HandleWrapperProps {
     id: string;
@@ -134,17 +86,6 @@ export const HandleWrapper = memo(
             return null;
         }, [connectedEdge, functionDefinitions, typeState, getNode]);
 
-        // A conic gradient that uses all handle colors to give an even distribution of colors
-        const handleColorString = handleColors
-            .map((color, index) => {
-                const percent = index / handleColors.length;
-                const nextPercent = (index + 1) / handleColors.length;
-                return `${color} ${percent * 100}% ${nextPercent * 100}%`;
-            })
-            .join(', ');
-        const handleGradient = `conic-gradient(from 90deg, ${handleColorString})`;
-        const connectedColor = 'var(--connection-color)';
-
         return (
             <HStack h="full">
                 <Center
@@ -152,45 +93,15 @@ export const HandleWrapper = memo(
                     left="-6px"
                     position="absolute"
                 >
-                    <Div
-                        _before={{
-                            content: '" "',
-                            position: 'absolute',
-                            top: '50%',
-                            left: '50%',
-                            height: '30px',
-                            width: '45px',
-                            cursor: 'crosshair',
-                            transform: 'translate(-50%, -50%)',
-                            borderRadius: '100%',
-                        }}
-                        _hover={{
-                            width: '22px',
-                            height: '22px',
-                            marginLeft: '-3px',
-                            opacity: validity.isValid ? 1 : 0,
-                        }}
-                        as={LeftHandle}
-                        className="input-handle"
+                    <Handle
+                        connectedColor={
+                            isConnected ? parentTypeColor ?? handleColors[0] : undefined
+                        }
+                        handleColors={handleColors}
                         id={targetHandle}
                         isValidConnection={isValidConnectionForRf}
-                        sx={{
-                            width: '16px',
-                            height: '16px',
-                            borderWidth: isConnected ? '2px' : '0px',
-                            borderColor: parentTypeColor ?? 'none',
-                            transition: '0.15s ease-in-out',
-                            ...(handleColors.length > 1
-                                ? { background: isConnected ? connectedColor : handleGradient }
-                                : {}),
-                            backgroundColor: isConnected ? connectedColor : handleColors[0],
-                            boxShadow: '2px 2px 2px #00000014',
-                            filter: validity.isValid ? undefined : 'grayscale(100%)',
-                            opacity: validity.isValid ? 1 : 0.3,
-                            position: 'relative',
-                        }}
+                        type="input"
                         validity={validity}
-                        onContextMenu={noContextMenu}
                     />
                 </Center>
                 {children}
